refactor(middleware): added global themer instance for quick usage in components
The module exports an instance of `react-themer` with `react-themer-jss` added as a middleware as

default export for quick usage in components. This allows components to skip themer setup. The

module also exports a `createMiddleware` utility that allows to create a middleware to add to

`themer`.  Also added tests.

named export `create` is renamed to `createMiddleware`